+++ CHANGELOG.md
@@ -5,8 +5,8 @@
 ### Added
 - Support cost control configured in the Dynatrace UI.
 - It is possible to get the duration from an `Action`.
-- An `Action` can be cancelled.  
-  Cancelling an Action is similar to leaving it, without reporting it.
+- An `Action` can be canceled.  
+  Canceling an `Action` is similar to leaving it, without reporting it.
 - Requests to Dynatrace can be intercepted to add custom HTTP headers.
 
 ### Changed